benchmarking/requirements.txt
google-cloud-vision>=3.4.0
openai>=1.3.0
ollama>=0.1.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
fuzzy>=1.0.1
difflib-python>=0.1.0
